# swagger_test.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    ChangePasswordResp:
        properties:
            result:
                type: string
                x-go-name: Result
        type: object
        x-go-package: go-zrbc/view
    CommonError:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            err:
                type: string
                x-go-name: Err
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: go-zrbc/docs
...
    EditLimitResp:
        properties:
            result:
                type: string
                x-go-name: Result
        type: object
        x-go-package: go-zrbc/view
    GetAgentBalanceResp:
        properties:
            result:
                description: 代理商余额
                type: string
                x-go-name: Result
        type: object
        x-go-package: go-zrbc/view
...
    GetUserInfoResp:
        properties:
            data:
                $ref: '#/definitions/Member'
        type: object
        x-go-package: go-zrbc/view
    LogoutGameResp:
        properties:
            result:
                type: string
                x-go-name: Result
        type: object
        x-go-package: go-zrbc/view
    Member:
        properties:
            cash:
                description: 現金
                type: string
                x-go-name: Cash
            chips:
                description: 籌碼選擇
                type: string
                x-go-name: Chips
            conwinprompt:
                description: 連贏局數告警
                format: int64
                type: integer
                x-go-name: Conwinprompt
            currency:
                description: 幣別
                format: int64
                type: integer
                x-go-name: Currency
            follow1:
                description: 注關會員
                type: string
                x-go-name: Follow1
            follow2:
                description: 關注荷官
                type: string
                x-go-name: Follow2
            head:
                description: 頭像ID
                format: int64
                type: integer
                x-go-name: Head
            id:
                description: 主键
                format: int64
                type: integer
                x-go-name: ID
            identity:
                description: 0:信用,1:api
                format: int64
                type: integer
                x-go-name: Identity
            kickperiod:
                description: 踢出局數
                format: int64
                type: integer
                x-go-name: Kickperiod
            lineid:
                description: Line ID
                type: string
                x-go-name: Lineid
            lockmoney:
                description: 鎖定金額
                type: string
                x-go-name: Lockmoney
            mem005:
                format: int64
                type: integer
                x-go-name: Mem005
            mem006:
                format: int64
                type: integer
                x-go-name: Mem006
            mem007:
                format: int64
                type: integer
                x-go-name: Mem007
            mem008:
                format: int64
                type: integer
                x-go-name: Mem008
            mem009:
                format: int64
                type: integer
                x-go-name: Mem009
            mem010:
                format: int64
                type: integer
                x-go-name: Mem010
            mem011:
                format: int64
                type: integer
                x-go-name: Mem011
            mem012:
                format: int64
                type: integer
                x-go-name: Mem012
            mem013:
                format: int64
                type: integer
                x-go-name: Mem013
            mem014:
                type: string
                x-go-name: Mem014
            mem015:
                description: login_error
                format: int64
                type: integer
                x-go-name: Mem015
            mem016:
                description: enable
                type: string
                x-go-name: Mem016
            mem017:
                description: canbet
                type: string
                x-go-name: Mem017
            mem018:
                description: chg_pw
                type: string
                x-go-name: Mem018
            mem019:
                description: is_test
                type: string
                x-go-name: Mem019
            mem020:
                description: be_traded
                type: string
                x-go-name: Mem020
            mem021:
                format: int64
                type: integer
                x-go-name: Mem021
            mem022:
                description: 電話
                type: string
                x-go-name: Mem022
            mem022a:
                description: 電話簡碼
                format: int64
                type: integer
                x-go-name: Mem022a
            mem023:
                format: int64
                type: integer
                x-go-name: Mem023
            mem024:
                format: int64
                type: integer
                x-go-name: Mem024
            mem026:
                description: mem_risk
                format: int64
                type: integer
                x-go-name: Mem026
            mem028:
                description: 備註
                type: string
                x-go-name: Mem028
            money:
                description: 己出碼額度
                type: string
                x-go-name: Money
            onlineprompt:
                description: 上線提示
                type: string
                x-go-name: Onlineprompt
            opengame:
                description: 輸入要開啟的種類
                type: string
                x-go-name: Opengame
            profitprompt:
                description: 盈利告警
                format: int64
                type: integer
                x-go-name: Profitprompt
            red:
                description: 红包开关
                type: string
                x-go-name: Red
            singlebetprompt:
                description: 單注金額告警
                type: string
                x-go-name: Singlebetprompt
            site:
                type: string
                x-go-name: Site
            tip:
                type: string
                x-go-name: Tip
            type:
                description: 0:現金;1:信用;2:電投
                format: int64
                type: integer
                x-go-name: Type
            user:
                description: 用户名
                type: string
                x-go-name: User
            username:
                description: 用户名称
                type: string
                x-go-name: UserName
            wallet:
                type: string
                x-go-name: Wallet
            winlossprompt:
                description: 最大可輸可贏提示
                type: string
                x-go-name: Winlossprompt
        type: object
        x-go-package: go-zrbc/view
    MemberRegisterResp:
        properties:
            result:
                type: string
                x-go-name: Result
        type: object
        x-go-package: go-zrbc/view
...
    SigninGameResp:
        properties:
            result:
                type: string
                x-go-name: Result
        type: object
        x-go-package: go-zrbc/view
...
host: wys.test.zhanggao223.com
info:
    description: Documentation of our go-zrbc-api API.
    title: go-zrbc-api
    version: 1.0.0
paths:
    /v1/change_password:
        post:
            consumes:
                - multipart/form-data
            description: 修改密码
            operationId: ChangePassword
            parameters:
                - description: 代理商(aid)
                  in: formData
                  name: vendorId
                  type: string
                  x-go-name: VendorID
                - description: 代理商标识符
                  in: formData
                  name: signature
                  type: string
                  x-go-name: Signature
                - description: 帐号
                  in: formData
                  name: user
                  type: string
                  x-go-name: User
                - description: 新密码
                  in: formData
                  name: newpassword
                  type: string
                  x-go-name: NewPassword
                - description: 时间戳
                  format: int64
                  in: formData
                  name: timestamp
                  type: integer
                  x-go-name: Timestamp
                - description: 0:中文, 1:英文 (非必要)
                  format: int64
                  in: formData
                  name: syslang
                  type: integer
                  x-go-name: Syslang
            responses:
                "200":
                    description: ChangePasswordResp
                    schema:
                        $ref: '#/definitions/ChangePasswordResp'
                "500":
                    $ref: '#/responses/CommonError'
            tags:
                - api渠道接口
...
    /v1/edit_limit:
        post:
            consumes:
                - multipart/form-data
            description: 修改限额
            operationId: EditLimit
            parameters:
                - description: 代理商(aid)
                  in: formData
                  name: vendorId
                  type: string
                  x-go-name: VendorID
                - description: 代理商标识符
                  in: formData
                  name: signature
                  type: string
                  x-go-name: Signature
                - description: 帐号
                  in: formData
                  name: user
                  type: string
                  x-go-name: User
                - description: 限制类型 (非必要)例:2,5,9,14,48,107,111,131 (非必要)
                  in: formData
                  name: limitType
                  type: string
                  x-go-name: LimitType
                - description: 最大赢（非必要）
                  format: int64
                  in: formData
                  name: maxwin
                  type: integer
                  x-go-name: Maxwin
                - description: 最大输（非必要）
                  format: int64
                  in: formData
                  name: maxlose
                  type: integer
                  x-go-name: Maxlose
                - description: 重置时间
                  format: int64
                  in: formData
                  name: reset
                  type: integer
                  x-go-name: Reset
                - description: 时间戳
                  format: int64
                  in: formData
                  name: timestamp
                  type: integer
                  x-go-name: Timestamp
                - description: 0:中文, 1:英文 (非必要)
                  format: int64
                  in: formData
                  name: syslang
                  type: integer
                  x-go-name: Syslang
            responses:
                "200":
                    description: EditLimitResp
                    schema:
                        $ref: '#/definitions/EditLimitResp'
                "500":
                    $ref: '#/responses/CommonError'
            tags:
                - api渠道接口
    /v1/get_agent_balance:
        post:
            consumes:
                - multipart/form-data
            description: 获取代理商余额
            operationId: GetAgentBalance
            parameters:
                - description: 代理商(aid)
                  in: formData
                  name: vendorId
                  type: string
                  x-go-name: VendorID
                - description: 代理商标识符
                  in: formData
                  name: signature
                  type: string
                  x-go-name: Signature
                - description: 时间戳
                  format: int64
                  in: formData
                  name: timestamp
                  type: integer
                  x-go-name: Timestamp
                - description: 0:中文, 1:英文 (非必要)
                  format: int64
                  in: formData
                  name: syslang
                  type: integer
                  x-go-name: Syslang
            responses:
                "200":
                    description: GetAgentBalanceResp
                    schema:
                        $ref: '#/definitions/GetAgentBalanceResp'
                "500":
                    $ref: '#/responses/CommonError'
            tags:
                - api渠道接口
...
    /v1/logout_game:
        post:
            consumes:
                - multipart/form-data
            description: 登出游戏
            operationId: LogoutGame
            parameters:
                - description: 代理商(aid)
                  in: formData
                  name: vendorId
                  type: string
                  x-go-name: VendorID
                - description: 代理商标识符
                  in: formData
                  name: signature
                  type: string
                  x-go-name: Signature
                - description: 帐号
                  in: formData
                  name: user
                  type: string
                  x-go-name: User
                - description: 时间戳
                  format: int64
                  in: formData
                  name: timestamp
                  type: integer
                  x-go-name: Timestamp
                - description: 0:中文, 1:英文 (非必要)
                  format: int64
                  in: formData
                  name: syslang
                  type: integer
                  x-go-name: Syslang
            responses:
                "200":
                    description: LogoutGameResp
                    schema:
                        $ref: '#/definitions/LogoutGameResp'
                "500":
                    $ref: '#/responses/CommonError'
            tags:
                - api渠道接口
    /v1/member_register:
        post:
            consumes:
                - multipart/form-data
            description: 注册用户
            operationId: MemberRegister
            parameters:
                - description: 代理商(aid)
                  in: formData
                  name: vendorId
                  type: string
                  x-go-name: VendorID
                - description: 代理商标识符
                  in: formData
                  name: signature
                  type: string
                  x-go-name: Signature
                - description: 帐号
                  in: formData
                  name: user
                  type: string
                  x-go-name: User
                - in: formData
                  name: password
                  type: string
                  x-go-name: Password
                - description: 用户名称
                  in: formData
                  name: username
                  type: string
                  x-go-name: Username
                - description: 用户类型（非必要）
                  format: int64
                  in: formData
                  name: profile
                  type: integer
                  x-go-name: Profile
                - description: 最大赢（非必要）
                  format: int64
                  in: formData
                  name: maxwin
                  type: integer
                  x-go-name: Maxwin
                - description: 最大输（非必要）
                  format: int64
                  in: formData
                  name: maxlose
                  type: integer
                  x-go-name: Maxlose
                - description: 备注（非必要）
                  in: formData
                  name: mark
                  type: string
                  x-go-name: Mark
                - description: '會員退水是否歸零   0為: 不歸零 1為: 歸零  (非必要)'
                  format: int64
                  in: formData
                  name: rakeback
                  type: integer
                  x-go-name: Rakeback
                - description: |-
                    限制类型 (非必要)例:2,5,9,14,48,107,111,131 (非必要)
                    (此为风险控管的一环，请经营者多善利用)
                    http://wmapi.a45.me/Limit.html
                  in: formData
                  name: limitType
                  type: string
                  x-go-name: LimitType
                - description: |-
                    籌碼選擇会员筹码 使用逗号隔开，可填入5-10组 (非必要)
                    若没带入，会员筹码预设为 10,20,50,100,500,1000,5000,10000,20000,50000
                    可用筹码种类：1, 5, 10, 20, 50, 100, 500, 1000, 5000, 10000, 20000, 50000, 100000, 200000, 1000000, 5000000, 10000000, 20000000, 50000000, 10000000, 20000000, 50000000, 100000000
                  in: formData
                  name: chip
                  type: string
                  x-go-name: Chips
                - description: 提出局数（非必要）
                  in: formData
                  name: kickperiod
                  type: string
                  x-go-name: Kickperiod
                - description: 时间戳
                  format: int64
                  in: formData
                  name: timestamp
                  type: integer
                  x-go-name: Timestamp
                - description: 0:中文, 1:英文 (非必要)
                  format: int64
                  in: formData
                  name: syslang
                  type: integer
                  x-go-name: Syslang
            responses:
                "200":
                    description: MemberRegisterResp
                    schema:
                        $ref: '#/definitions/MemberRegisterResp'
                "500":
                    $ref: '#/responses/CommonError'
            tags:
                - api渠道接口
    /v1/signin_game:
        post:
            consumes:
                - multipart/form-data
            description: 开游戏
            operationId: SigninGame
            parameters:
                - description: 代理商(aid)
                  in: formData
                  name: vendorId
                  type: string
                  x-go-name: VendorID
                - description: 代理商标识符
                  in: formData
                  name: signature
                  type: string
                  x-go-name: Signature
                - description: 帐号
                  in: formData
                  name: user
                  type: string
                  x-go-name: User
                - description: 设备
                  in: formData
                  name: device
                  type: string
                  x-go-name: Device
                - description: |-
                    语言 使用语言 0或空值 为简体中文
                    1为英文
                    2为泰文
                    3为越文
                    4为日文
                    5为韩文
                    6为印度文
                    7为马来西亚文
                    8为印尼文
                    9为繁体中文
                    10为西文
                    11为俄文
                  in: formData
                  name: lang
                  type: string
                  x-go-name: Lang
                - description: 试玩（非必要）
                  in: formData
                  name: isTest
                  type: boolean
                  x-go-name: IsTest
                - description: |-
                    游戏模式（非必要）onlybac 略过大厅直接进入百家乐；onlydgtg 略过大厅直接进入龙虎；onlyrou 略过大厅直接进入轮盘；onlysicbo 略过大厅直接进入骰宝；onlyniuniu 略过大厅直接进入牛牛；onlyfantan 略过大厅直接进入番摊；onlysedie 略过大厅直接进入色碟；下方數值務必帶入參數tableid
                    onlybactable 略过大厅直接进入百家乐；onlydgtgtable 略过大厅直接进入龙虎；onlyroutable 略过大厅直接进入轮盘；onlysicbotable 略过大厅直接进入骰宝；onlyniuniutable 略过大厅直接进入牛牛；onlyfantantable 略过大厅直接进入番摊；onlysedietable 略过大厅直接进入色碟
                  in: formData
                  name: mode
                  type: string
                  x-go-name: Mode
                - description: |-
                    桌子IDmode参数带入以下数值为必填
                    onlybactable
                    onlydgtgtable
                    onlyroutable
                    onlysicbotable
                    onlyniuniutable
                    onlyfantantable
                    onlysedietable
                    否则非必要
                  in: formData
                  name: tableid
                  type: string
                  x-go-name: TableID
                - description: 站点
                  in: formData
                  name: site
                  type: string
                  x-go-name: Site
                - in: formData
                  name: password
                  type: string
                  x-go-name: Password
                - description: 游戏类型
                  in: formData
                  name: gameType
                  type: string
                  x-go-name: GameType
                - description: 宽度
                  in: formData
                  name: width
                  type: string
                  x-go-name: Width
                - description: 返回URL
                  in: formData
                  name: returnurl
                  type: string
                  x-go-name: ReturnURL
                - description: 1:iframe嵌入 (非必要)
                  in: formData
                  name: size
                  type: string
                  x-go-name: Size
                - description: 用户ID
                  format: int64
                  in: formData
                  name: ui
                  type: integer
                  x-go-name: UI
                - description: 静音
                  in: formData
                  name: mute
                  type: string
                  x-go-name: Mute
                - description: 视频
                  in: formData
                  name: video
                  type: string
                  x-go-name: Video
                - description: 0:中文, 1:英文 (非必要)
                  format: int64
                  in: formData
                  name: syslang
                  type: integer
                  x-go-name: Syslang
                - description: 籌碼選擇
                  in: query
                  name: chip
                  type: string
                  x-go-name: Chips
                - description: 时间戳
                  format: int64
                  in: formData
                  name: timestamp
                  type: integer
                  x-go-name: Timestamp
            responses:
                "200":
                    description: SigninGameResp
                    schema:
                        $ref: '#/definitions/SigninGameResp'
                "500":
                    $ref: '#/responses/CommonError'
            tags:
                - api渠道接口
...
    /v1/user_info:
        get:
            description: 获取用户信息
            operationId: GetUserInfo
            parameters:
                - in: header
                  name: Authorization
                  type: string
                  x-go-name: Token
            responses:
                "200":
                    description: GetUserInfoResp
                    schema:
                        $ref: '#/definitions/GetUserInfoResp'
                "500":
                    $ref: '#/responses/CommonError'
            tags:
                - 用户接口
produces:
    - application/json
responses:
...
    CommonError:
        description: Means the server has error
        schema:
            $ref: '#/definitions/CommonError'
...
schemes:
    - https
securityDefinitions:
    basic:
        type: basic
swagger: "2.0"